Make IntermediateToolingModelProvider Build scoped
